Move the editor karaoke sprite text into the matched namespace.
