+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Edit/Blueprints/TimeTagSelectionBlueprint.cs
@@ -6,7 +6,6 @@
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Game.Rulesets.Edit;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osuTK;
