+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Edit/Carets/DrawableTimeTagRecordCaret.cs
@@ -8,7 +8,6 @@ using osu.Framework.Graphics;
 using osu.Game.Graphics;
 using osu.Game.Rulesets.Karaoke.Edit.Components.Sprites;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
 using osu.Game.Rulesets.Karaoke.Utils;
 using osuTK;
