+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Edit/SingleLyricEditor.cs
@@ -4,7 +4,6 @@
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Screens.Edit;
